Programatic tests of path interpolation - fixed problems in parenthesis in math expressions that we resulting in the incorrect angles being generated when the interpolated positions were converted back to

+++ frc 2019/src/main/java/frc/robot/subsystems/ArmDriveSrx.java
@@ -54,7 +54,7 @@ public class ArmDriveSrx extends Subsystem implements IUseArm {
     }
 
     @Override
-    public double getBucketAngle() {
+    public int getBucketAngle() {
         return 0;
     }
