+++ frc 2019/src/main/java/frc/robot/subsystems/IUseArm.java
@@ -26,9 +26,9 @@ public interface IUseArm {
     /**
      * Get the angle of the bucket.
      *
-     * @return (double) The position angle of the bucket
+     * @return (int) The position angle of the bucket
      */
-    double getBucketAngle();
+    int getBucketAngle();
 
     /**
      * Set the arm motor power (or speed, depending on the implementation)for the lower arm.
